+++ PharmInventory/Scripts/9988-proc_Model22Insert.sql
@@ -2,7 +2,7 @@ CREATE PROCEDURE [proc_Model22Insert]
 (
 	@ID int = NULL output,
 	@IssueDocID int,
-	@PackQty int = NULL,
+	@PackQty decimal(18,0) = NULL,
 	@QtyPerPack decimal(18,0) = NULL,
 	@PackPrice decimal(16,4) = NULL,
 	@TotalPrice decimal(16,4) = NULL,
